+++ src/main/java/com/bitbank/repositories/ERole.java
@@ -1,0 +1,5 @@
+package com.bitbank.repositories;
+
+public enum ERole {
+	ROLE_ADMIN, ROLE_CUSTOMER
+}