+++ src/test/java/com/bitbank/controllers/AuthControllerTest.java
@@ -1,13 +1,19 @@
 package com.bitbank.controllers;
 
 import static org.hamcrest.CoreMatchers.is;
+import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.when;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
+import java.util.HashSet;
+import java.util.Optional;
+import java.util.Set;
 import java.util.stream.Stream;
 
+import org.hamcrest.Matchers;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.junit.jupiter.params.ParameterizedTest;
@@ -27,7 +33,12 @@ import org.springframework.test.web.servlet.MockMvc;
 
 import com.bitbank.config.AuthEntryPointJwt;
 import com.bitbank.config.AuthTokenFilter;
+import com.bitbank.entities.RolesEntity;
 import com.bitbank.entities.UsersEntity;
+import com.bitbank.exceptions.RoleOrAuthorityNotFoundException;
+import com.bitbank.repositories.ERole;
+import com.bitbank.repositories.RolesRepository;
+import com.bitbank.services.RolesService;
 import com.bitbank.services.UserDetailsImpl;
 import com.bitbank.services.UserDetailsServiceImpl;
 import com.bitbank.utils.JwtUtils;
@@ -65,27 +76,85 @@ class AuthControllerTest {
 	@MockBean
 	SecurityContext securityContext;
 
+	@MockBean
+	private RolesService rolesService;
+
+	@MockBean
+	private RolesRepository rolesRepository;
+
+	/**
+	 * Generate an authority before every test
+	 * 
+	 * @return
+	 */
+	@BeforeEach
+	private void buildAdminRoleAuthority() {
+		// ADMIN - Initialize the set
+		Set<RolesEntity> adminRole = new HashSet<>();
+		// ADMIN - Generate the entity
+		RolesEntity adminRolesEntity = validRolesEntity(1L, ERole.valueOf("ROLE_ADMIN"));
+		// ADMIN - Add the entity to the set
+		adminRole.add(adminRolesEntity);
+		this.adminRole = adminRole;
+	}
+
+	@BeforeEach
+	private void buildCustomerRoleAuthority() {
+		// CUSTOMER - Initialize the set
+		Set<RolesEntity> customerRole = new HashSet<>();
+		// CUSTOMER - Generate the entity
+		RolesEntity customerRolesEntity = validRolesEntity(1L, ERole.valueOf("ROLE_CUSTOMER"));
+		// CUSTOMER - Add the entity to the set
+		customerRole.add(customerRolesEntity);
+		this.customerRole = customerRole;
+	}
+
+	/**
+	 * The autority to use
+	 * 
+	 * @throws Exception
+	 */
+	private Set adminRole;
+	private Set customerRole;
+
+	private Set getAdminRole() {
+		return this.adminRole;
+	}
+
+	private Set getCustomerRole() {
+		return this.customerRole;
+	}
+
 	@Test
+	@WithMockUser(username = "username", authorities = { "ROLE_ADMIN" })
 	void canRegisterNewUser() throws Exception {
 
-		var userToSave = validUserEntity("username", "password");
-		var savedUser = validUserEntity("username", "a1.b2.c3");
+		// Create a valid user
+		var userToSave = validUserEntity("username", "password", this.getCustomerRole());
+		var savedUser = validUserEntity("username", "a1.b2.c3", this.getCustomerRole());
 
+		// Create valid role entity
+		RolesEntity rolesEntity = validRolesEntity(1L, ERole.valueOf("ROLE_CUSTOMER"));
+		// Create an optional result
+		RolesEntity result = Optional.of(rolesEntity).get();
+
+		// Mock some methods
 		when(userDetailsServiceImpl.post(userToSave)).thenReturn(savedUser);
+		when(rolesService.show(ERole.valueOf("ROLE_CUSTOMER"))).thenReturn(result);
 
 		mvc.perform(post("/api/v1/auth/register/").contentType(MediaType.APPLICATION_JSON)
 				.content(objectMapper.writeValueAsBytes(userToSave))).andExpect(status().isCreated());
 	}
 
-	@WithMockUser("username")
 	@Test
+	@WithMockUser(username = "username", authorities = { "ROLE_ADMIN" })
 	void canLogin() throws Exception {
 
 		// Add security context
 		SecurityContextHolder.setContext(securityContext);
 
 		// Create a new usersentity to deal with authentication
-		UsersEntity usersEntity = new UsersEntity(1L, "username", "password");
+		UsersEntity usersEntity = new UsersEntity(1L, "username", "password", this.getAdminRole());
 		// Build an user from usersEntity
 		UserDetailsImpl user = UserDetailsImpl.build(usersEntity);
 
@@ -94,18 +163,16 @@ class AuthControllerTest {
 		when(authentication.getPrincipal()).thenReturn(user);
 
 		String token = jwtUtils.generateJwtToken(authentication);
-		
-		System.out.println(token);
-
 		Long expiryDate = jwtUtils.getExpiryDateFromJwtToken(token);
 
 		// Create an usersEntity to pass to the login
-		var userToLogin = validUserEntity("username", "password");
+		UsersEntity userToLogin = validUserEntity("username", "password", this.getAdminRole());
 
 		mvc.perform(post("/api/v1/auth/login/").contentType(MediaType.APPLICATION_JSON)
 				.content(objectMapper.writeValueAsBytes(userToLogin))).andExpect(status().isOk())
 				.andExpect(jsonPath("$.access_token", is(token)))
-				.andExpect(jsonPath("$.expiry_at", is(expiryDate.toString())));
+				.andExpect(jsonPath("$.expiry_at", is(expiryDate.toString())))
+				.andExpect(jsonPath("$.roles", Matchers.empty()));
 	}
 
 	/**
@@ -123,15 +190,15 @@ class AuthControllerTest {
 
 	}
 
-	@WithMockUser("username")
 	@Test
+	@WithMockUser(username = "username", authorities = { "ROLE_CUSTOMER" })
 	void testCanRefreshToken() throws Exception {
 
 		// Add the Security Context
 		SecurityContextHolder.setContext(securityContext);
 
 		// Create an usersEntity to build by userDetailsImpl
-		UsersEntity usersEntity = new UsersEntity(1L, "username", "password");
+		UsersEntity usersEntity = new UsersEntity(1L, "username", "password", this.getAdminRole());
 		UserDetailsImpl user = UserDetailsImpl.build(usersEntity);
 
 		// Mock some method...
@@ -173,16 +240,52 @@ class AuthControllerTest {
 	 * @return
 	 */
 	private static Stream<UsersEntity> getInvalidUsers() {
+		// ADMIN - Initialize the set
+		Set<RolesEntity> adminRole = new HashSet<>();
+		// ADMIN - Generate the entity
+		RolesEntity adminRolesEntity = validRolesEntity(1L, ERole.valueOf("ROLE_ADMIN"));
+		// ADMIN - Add the entity to the set
+		adminRole.add(adminRolesEntity);
 		// Create a valid entity
-		var validUsersEntity = validUserEntity("username", "password");
+		var validUsersEntity = validUserEntity("username", "password", adminRole);
 
 		return Stream.of(validUsersEntity.toBuilder().username(null).build(),
 				validUsersEntity.toBuilder().username("").build(), validUsersEntity.toBuilder().password(null).build(),
 				validUsersEntity.toBuilder().password("").build());
 	}
 
-	private static UsersEntity validUserEntity(String username, String password) {
-		return UsersEntity.builder().username(username).password(password).build();
+	/**
+	 * Test RoleOrAuthorityNotFoundException
+	 * 
+	 */
+	@Test
+	void canCatchRoleOrAuthorityNotFoundException() throws Exception {
+
+		// Create a valid user
+		var userToSave = validUserEntity("username", "password", this.getAdminRole());
+
+		when(rolesService.show(any())).thenThrow(new RoleOrAuthorityNotFoundException("cannot find role ADMIN_ROLE"));
+
+		mvc.perform(post("/api/v1/auth/register/").contentType(MediaType.APPLICATION_JSON)
+				.content(objectMapper.writeValueAsBytes(userToSave))).andExpect(status().isBadRequest());
+	}
+
+	/**
+	 * Generate a user entity
+	 * 
+	 * @param username
+	 * @param password
+	 * @return
+	 */
+	private static UsersEntity validUserEntity(String username, String password, Set roles) {
+		return UsersEntity.builder().username(username).password(password).rolesEntity(roles).build();
+	}
+
+	/**
+	 * Generate a RolesEntity
+	 */
+	private static RolesEntity validRolesEntity(Long id, ERole role) {
+		return RolesEntity.builder().id(id).role(role).build();
 	}
 
 }