+++ src/test/java/com/bitbank/controllers/RegisterControllerTest.java
@@ -15,6 +15,7 @@ import org.springframework.http.MediaType;
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContext;
+import org.springframework.security.test.context.support.WithMockUser;
 import org.springframework.test.context.TestPropertySource;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
 import org.springframework.test.web.servlet.MockMvc;
@@ -22,6 +23,7 @@ import org.springframework.test.web.servlet.MockMvc;
 import com.bitbank.config.AuthEntryPointJwt;
 import com.bitbank.config.AuthTokenFilter;
 import com.bitbank.entities.UsersEntity;
+import com.bitbank.services.RolesService;
 import com.bitbank.services.UserDetailsServiceImpl;
 import com.bitbank.utils.JwtUtils;
 import com.fasterxml.jackson.databind.ObjectMapper;
@@ -58,7 +60,11 @@ class RegisterControllerTest {
 	@MockBean
 	SecurityContext securityContext;
 
+	@MockBean
+	private RolesService rolesService;
+
 	@Test
+	@WithMockUser(username = "username", authorities = { "ROLE_ADMIN" })
 	void cannotRegisterUserWhenRegistrationsAreDisabled() throws Exception {
 
 		var userToSave = validUserEntity("username", "password");