+++ src/test/java/com/bitbank/utils/JwtUtilsTest.java
@@ -1,10 +1,15 @@
 package com.bitbank.utils;
 
+import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.mockito.Mockito.when;
+
+import java.util.HashSet;
+import java.util.List;
+import java.util.Set;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -17,7 +22,9 @@ import org.springframework.security.core.context.SecurityContext;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.test.context.TestPropertySource;
 
+import com.bitbank.entities.RolesEntity;
 import com.bitbank.entities.UsersEntity;
+import com.bitbank.repositories.ERole;
 import com.bitbank.services.UserDetailsImpl;
 
 @SpringBootTest
@@ -45,7 +52,15 @@ class JwtUtilsTest {
 	@Test
 	void canGenerateToken() {
 
-		UsersEntity usersEntity = new UsersEntity(1L, "username", "password");
+		// ADMIN - Initialize the set
+		Set<RolesEntity> adminRole = new HashSet<>();
+		// ADMIN - Generate the entity
+		RolesEntity adminRolesEntity = validRolesEntity(1L, ERole.valueOf("ROLE_ADMIN"));
+		// ADMIN - Add the entity to the set
+		adminRole.add(adminRolesEntity);
+
+		UsersEntity usersEntity = new UsersEntity(1L, "username", "password", adminRole);
+
 		UserDetailsImpl user = UserDetailsImpl.build(usersEntity);
 
 		when(timeSource.getCurrentTimeMillis()).thenReturn(System.currentTimeMillis());
@@ -58,7 +73,15 @@ class JwtUtilsTest {
 
 	@Test
 	void canGetUserNameFromJwtToken() {
-		UsersEntity usersEntity = new UsersEntity(1L, "username", "password");
+
+		// ADMIN - Initialize the set
+		Set<RolesEntity> adminRole = new HashSet<>();
+		// ADMIN - Generate the entity
+		RolesEntity adminRolesEntity = validRolesEntity(1L, ERole.valueOf("ROLE_ADMIN"));
+		// ADMIN - Add the entity to the set
+		adminRole.add(adminRolesEntity);
+
+		UsersEntity usersEntity = new UsersEntity(1L, "username", "password", adminRole);
 		UserDetailsImpl user = UserDetailsImpl.build(usersEntity);
 
 		when(timeSource.getCurrentTimeMillis()).thenReturn(System.currentTimeMillis());
@@ -73,7 +96,15 @@ class JwtUtilsTest {
 
 	@Test
 	void canValidateToken() {
-		UsersEntity usersEntity = new UsersEntity(1L, "username", "password");
+
+		// ADMIN - Initialize the set
+		Set<RolesEntity> adminRole = new HashSet<>();
+		// ADMIN - Generate the entity
+		RolesEntity adminRolesEntity = validRolesEntity(1L, ERole.valueOf("ROLE_ADMIN"));
+		// ADMIN - Add the entity to the set
+		adminRole.add(adminRolesEntity);
+
+		UsersEntity usersEntity = new UsersEntity(1L, "username", "password", adminRole);
 		UserDetailsImpl user = UserDetailsImpl.build(usersEntity);
 
 		when(timeSource.getCurrentTimeMillis()).thenReturn(System.currentTimeMillis());
@@ -94,7 +125,15 @@ class JwtUtilsTest {
 
 	@Test
 	void canCheckTokenIsExpired() {
-		UsersEntity usersEntity = new UsersEntity(1L, "username", "password");
+
+		// ADMIN - Initialize the set
+		Set<RolesEntity> adminRole = new HashSet<>();
+		// ADMIN - Generate the entity
+		RolesEntity adminRolesEntity = validRolesEntity(1L, ERole.valueOf("ROLE_ADMIN"));
+		// ADMIN - Add the entity to the set
+		adminRole.add(adminRolesEntity);
+
+		UsersEntity usersEntity = new UsersEntity(1L, "username", "password", adminRole);
 		UserDetailsImpl user = UserDetailsImpl.build(usersEntity);
 
 		// 1262344822000 == 01 Jan 2010 11:20:22 UTC
@@ -107,10 +146,22 @@ class JwtUtilsTest {
 		assertFalse(jwtUtils.validateJwtToken(token));
 	}
 
+	/**
+	 * Test can generate a token from itself.
+	 * 
+	 * Useful to refresh token itself.
+	 */
 	@Test
 	void canGenerateTokenFromTokenItself() {
 
-		UsersEntity usersEntity = new UsersEntity(1L, "username", "password");
+		// ADMIN - Initialize the set
+		Set<RolesEntity> adminRole = new HashSet<>();
+		// ADMIN - Generate the entity
+		RolesEntity adminRolesEntity = validRolesEntity(1L, ERole.valueOf("ROLE_ADMIN"));
+		// ADMIN - Add the entity to the set
+		adminRole.add(adminRolesEntity);
+
+		UsersEntity usersEntity = new UsersEntity(1L, "username", "password", adminRole);
 		UserDetailsImpl user = UserDetailsImpl.build(usersEntity);
 
 		when(timeSource.getCurrentTimeMillis()).thenReturn(System.currentTimeMillis());
@@ -137,7 +188,15 @@ class JwtUtilsTest {
 		when(securityContext.getAuthentication()).thenReturn(authentication);
 
 		// Create an usersEntitiy
-		UsersEntity usersEntity = new UsersEntity(1L, "username", "password");
+
+		// ADMIN - Initialize the set
+		Set<RolesEntity> adminRole = new HashSet<>();
+		// ADMIN - Generate the entity
+		RolesEntity adminRolesEntity = validRolesEntity(1L, ERole.valueOf("ROLE_ADMIN"));
+		// ADMIN - Add the entity to the set
+		adminRole.add(adminRolesEntity);
+
+		UsersEntity usersEntity = new UsersEntity(1L, "username", "password", adminRole);
 		// Build the entity to return from getPrincipal
 		UserDetailsImpl user = UserDetailsImpl.build(usersEntity);
 		when(authentication.getPrincipal()).thenReturn(user);
@@ -151,5 +210,44 @@ class JwtUtilsTest {
 		// Finally, assert equals. Accept a small clock shift
 		Long expectedExpiryDate = currentTimeMillis + Long.parseLong(jwtExpirationMs);
 		assertEquals(expectedExpiryDate / 10000, expiryDate / 10000);
+	}
+
+	@Test
+	void getGetRolesFromAuthentication() {
+
+		// Set the security context
+		SecurityContextHolder.setContext(securityContext);
+
+		// Mock the time at.. now
+		Long currentTimeMillis = System.currentTimeMillis();
+
+		// Mock the methods
+		when(timeSource.getCurrentTimeMillis()).thenReturn(currentTimeMillis);
+		when(securityContext.getAuthentication()).thenReturn(authentication);
+
+		// Create an usersEntitiy
+
+		// ADMIN - Initialize the set
+		Set<RolesEntity> adminRole = new HashSet<>();
+		// ADMIN - Generate the entity
+		RolesEntity adminRolesEntity = validRolesEntity(1L, ERole.valueOf("ROLE_ADMIN"));
+		// ADMIN - Add the entity to the set
+		adminRole.add(adminRolesEntity);
+
+		UsersEntity usersEntity = new UsersEntity(1L, "username", "password", adminRole);
+		// Build the entity to return from getPrincipal
+		UserDetailsImpl user = UserDetailsImpl.build(usersEntity);
+		when(authentication.getPrincipal()).thenReturn(user);
+
+		List<String> roles = jwtUtils.getAuthorities(authentication);
+
+		assertThat(roles).contains("ROLE_ADMIN").doesNotContain("ROLE_CUSTOMER");
+	}
+
+	/**
+	 * Generate a RolesEntity
+	 */
+	private static RolesEntity validRolesEntity(Long id, ERole role) {
+		return RolesEntity.builder().id(id).role(role).build();
 	}
 }